Move files around for packaging
